Clean up: removed system outs, renamed packages and commented out some code
