--- src/main/java/sod/eastonone/elasticsearch/EastonOneElasticSearchApplication.java
@@ -1,13 +1,0 @@
-package sod.eastonone.elasticsearch;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EastonOneElasticSearchApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EastonOneElasticSearchApplication.class, args);
-	}
-
-}
