--- src/main/java/sod/eastonone/elasticsearch/es/config/ESRestClient.java
@@ -1,60 +1,0 @@
-package sod.eastonone.elasticsearch.es.config;
-
-import co.elastic.clients.elasticsearch.ElasticsearchClient;
-import co.elastic.clients.json.jackson.JacksonJsonpMapper;
-import co.elastic.clients.transport.ElasticsearchTransport;
-import co.elastic.clients.transport.rest_client.RestClientTransport;
-import lombok.Getter;
-import lombok.Setter;
-import org.apache.http.HttpHost;
-import org.apache.http.auth.AuthScope;
-import org.apache.http.auth.UsernamePasswordCredentials;
-import org.apache.http.client.CredentialsProvider;
-import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.elasticsearch.client.RestClient;
-import org.elasticsearch.client.RestClientBuilder;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-
-@Configuration
-@ConfigurationProperties("es")
-//@Getter 
-//@Setter //not working, was working then suddenly hostName and port were null
-public class ESRestClient {
-
-    private String hostName;
-    private int port;
-    private String username;
-    private String password;
-
-    @Bean
-    public ElasticsearchClient getElasticSearchClient() {
-
-//        final CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
-//        credentialsProvider.setCredentials(AuthScope.ANY,
-//                new UsernamePasswordCredentials(username, password));
-    	System.out.println("hostName: " + hostName + " port: " + port);
-        RestClientBuilder builder = RestClient.builder(new HttpHost(hostName, port));
-//                .setHttpClientConfigCallback(httpClientBuilder ->
-//                        httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider));
-
-        // Create the low-level client
-        RestClient restClient = builder.build();
-
-        // Create the transport with a Jackson mapper
-        ElasticsearchTransport transport = new RestClientTransport(restClient, new JacksonJsonpMapper());
-
-        // And create the API client
-        return new ElasticsearchClient(transport);
-    }
-
-	public void setHostName(String hostName) {
-		this.hostName = hostName;
-	}
-
-	public void setPort(int port) {
-		this.port = port;
-	}
-}
