--- src/main/java/sod/eastonone/elasticsearch/es/exception/RecordNotFoundException.java
@@ -1,7 +1,0 @@
-package sod.eastonone.elasticsearch.es.exception;
-
-public class RecordNotFoundException extends RuntimeException{
-    public RecordNotFoundException(String message) {
-        super(message);
-    }
-}
