+++ src/main/java/sod/eastonone/music/EastonOneElasticSearchApplication.java
@@ -1,0 +1,13 @@
+package sod.eastonone.music;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EastonOneElasticSearchApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EastonOneElasticSearchApplication.class, args);
+	}
+
+}
