+++ src/main/java/sod/eastonone/music/dao/repository/SodSongRepository.java
@@ -1,0 +1,10 @@
+package sod.eastonone.music.dao.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import sod.eastonone.music.dao.entity.SodSong;
+
+@Repository
+public interface SodSongRepository extends JpaRepository<SodSong, Integer> {
+}
