+++ src/main/java/sod/eastonone/music/es/config/ESRestClient.java
@@ -1,0 +1,54 @@
+package sod.eastonone.music.es.config;
+
+import org.apache.http.HttpHost;
+import org.elasticsearch.client.RestClient;
+import org.elasticsearch.client.RestClientBuilder;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import co.elastic.clients.elasticsearch.ElasticsearchClient;
+import co.elastic.clients.json.jackson.JacksonJsonpMapper;
+import co.elastic.clients.transport.ElasticsearchTransport;
+import co.elastic.clients.transport.rest_client.RestClientTransport;
+
+
+@Configuration
+@ConfigurationProperties("es")
+//@Getter 
+//@Setter //not working, was working then suddenly hostName and port were null
+public class ESRestClient {
+
+    private String hostName;
+    private int port;
+    private String username;
+    private String password;
+
+    @Bean
+    public ElasticsearchClient getElasticSearchClient() {
+
+//        final CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+//        credentialsProvider.setCredentials(AuthScope.ANY,
+//                new UsernamePasswordCredentials(username, password));
+        RestClientBuilder builder = RestClient.builder(new HttpHost(hostName, port));
+//                .setHttpClientConfigCallback(httpClientBuilder ->
+//                        httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider));
+
+        // Create the low-level client
+        RestClient restClient = builder.build();
+
+        // Create the transport with a Jackson mapper
+        ElasticsearchTransport transport = new RestClientTransport(restClient, new JacksonJsonpMapper());
+
+        // And create the API client
+        return new ElasticsearchClient(transport);
+    }
+
+	public void setHostName(String hostName) {
+		this.hostName = hostName;
+	}
+
+	public void setPort(int port) {
+		this.port = port;
+	}
+}
