+++ src/main/java/sod/eastonone/music/es/exception/RecordNotFoundException.java
@@ -1,0 +1,7 @@
+package sod.eastonone.music.es.exception;
+
+public class RecordNotFoundException extends RuntimeException{
+    public RecordNotFoundException(String message) {
+        super(message);
+    }
+}
