+++ src/main/java/sod/eastonone/music/es/model/Employee.java
@@ -1,0 +1,76 @@
+package sod.eastonone.music.es.model;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class Employee {
+	
+//	***********************************************************************************************
+//	***********************************************************************************************
+//	**********************     Unused Class, kept as a reference       ****************************
+//	***  https://blogs.perficient.com/2022/08/22/elasticsearch-java-api-client-springboot/  *******
+//	******     https://github.com/sundharamurali/elasticsearch-springboot                 *********
+//	***********************************************************************************************
+	
+    private Long id;
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String gender;
+    private String jobTitle;
+    private String phone;
+    private Integer size;
+	public Long getId() {
+		return id;
+	}
+	public void setId(Long id) {
+		this.id = id;
+	}
+	public String getFirstName() {
+		return firstName;
+	}
+	public void setFirstName(String firstName) {
+		this.firstName = firstName;
+	}
+	public String getLastName() {
+		return lastName;
+	}
+	public void setLastName(String lastName) {
+		this.lastName = lastName;
+	}
+	public String getEmail() {
+		return email;
+	}
+	public void setEmail(String email) {
+		this.email = email;
+	}
+	public String getGender() {
+		return gender;
+	}
+	public void setGender(String gender) {
+		this.gender = gender;
+	}
+	public String getJobTitle() {
+		return jobTitle;
+	}
+	public void setJobTitle(String jobTitle) {
+		this.jobTitle = jobTitle;
+	}
+	public String getPhone() {
+		return phone;
+	}
+	public void setPhone(String phone) {
+		this.phone = phone;
+	}
+	public Integer getSize() {
+		return size;
+	}
+	public void setSize(Integer size) {
+		this.size = size;
+	}
+}
