--- src/test/java/sod/eastonone/elasticsearch/EastonOneElasticSearchApplicationTests.java
@@ -1,13 +1,0 @@
-package sod.eastonone.elasticsearch;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EastonOneElasticSearchApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
