+++ src/test/java/sod/eastonone/music/EastonOneMusicApplicationTests.java
@@ -1,0 +1,13 @@
+package sod.eastonone.music;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EastonOneMusicApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
